Initial implementation of the copy screen.
This is the initial work for the copy screen's user interface. All of
the UI elements are there and bound to the relevant data and messages.
It is not yet functional since the actual copying logic still needs to
be implemented.

As part of this work, did significant refactoring of the application
theme, added a logging library, and implemented the ability to configure
copy services.

The copy service module was also added. It currently just contains the
data definitions required by the UI and has no functionality.