Work-in-progress copy operation.
This adds the function that will handle the copy operation from start to
finish. It currently just has the validation and initial database
operations and is currently done on the same thread as the UI, but does
not do the actual copy.

This includes a number of added functionality to the database in order
to support the required database operations.