Update context.
Changed the application context to be a GObject instead of a native rust
data struct with a GObject wrapper since only the GObject version is
currently needed.

Updated the UI to create a context that gets passed down and bound from
the application window to the pages. Updated the copy page to use the
context's drive store instead of its own.

Made list view backgrounds transparent by default.