Add drive monitoring.
Imported the drive actor from another branch and implemented a future
for monitoring drive status and updating the UI accordingly.

As part of this, modified how the faux drives work so that they get the
information from JSON files instead of hard-coded data so that I can
modifiy the data while the application is running.